refactor: remove `direction` and `resizeMode`
