+++ apps/calendar/types/layout.d.ts
@@ -1,12 +1,3 @@
-import { Direction } from '@src/constants/layout';
-
-export interface PanelSize {
-  width: number | null;
-  height: number | null;
-  resizerWidth: number | null;
-  resizerHeight: number | null;
-}
-
 export interface RectSize {
   width: number;
   height: number;
@@ -21,7 +12,6 @@ export interface PanelRect extends RectSize {
 
 export interface PanelInfo {
   name: string;
-  direction?: Direction;
   overflowY?: boolean;
   overflowX?: boolean;
   show?: boolean;
@@ -42,5 +32,3 @@ export interface PanelInfo {
   resizerHeight?: number;
   resizerWidth?: number;
 }
-
-export type PanelElementRectMap = Record<string, PanelRect>;
